test: stop running legacy support unit tests inside Reborn test binaries
live_mission_helpers.rs and trace_runner.rs still carried inline
#[cfg(test)] mod tests blocks that support_unit_tests.rs was created to
absorb ("compile and run exactly once"). Because every Reborn binary
declares `mod support;` for trace_llm access, those 12 legacy tests ran
inside each of the 3 reborn_group_* bins under --features libsql — 36
redundant executions per CI run inside the Reborn partition.

Move both test blocks into support_unit_tests.rs (libsql-gated, matching
the existing test_rig_tests precedent) and widen the three private
trace_runner helpers they exercise to pub(crate). No test logic changed;
Reborn binaries now execute only Reborn tests.
